refactor(frontend): extract VrcMatchCard from event match list page
